+++ modules/string/src/cpp/StringCompare.cpp
@@ -43,7 +43,7 @@ compareString(std::wstring A, std::wstring B, bool bCaseSensitive, indexType len
         return compareString(strA, strB, false);
     }
     if (bCaseSensitive) {
-        bEq = (A.compare(B) == 0);
+        bEq = (A == B);
     } else {
         bEq = boost::iequals(A, B);
     }
@@ -99,8 +99,8 @@ StringCompare(ArrayOf A, ArrayOf B, bool bCaseSensitive, indexType len)
                 ArrayOf elementA = cellA[k];
                 ArrayOf elementB = cellB[k];
                 if (elementA.isRowVectorCharacterArray() && elementB.isRowVectorCharacterArray()) {
-                    Cp[k] = compareString(elementA.getContentAsWideString(),
-                        elementB.getContentAsWideString(), bCaseSensitive, len);
+                    Cp[k] = static_cast<Nelson::logical>(compareString(elementA.getContentAsWideString(),
+                        elementB.getContentAsWideString(), bCaseSensitive, len));
                 } else if (elementA.isCharacterArray() && elementB.isCharacterArray()) {
                     wstringVector s1 = elementA.getContentAsWideStringVector();
                     wstringVector s2 = elementB.getContentAsWideStringVector();
@@ -201,8 +201,8 @@ StringCompare(ArrayOf A, ArrayOf B, bool bCaseSensitive, indexType len)
             } else {
                 ArrayOf elementA = cellA[k];
                 if (elementA.isCharacterArray() && scalar2.isCharacterArray()) {
-                    Cp[k] = compareString(elementA.getContentAsWideString(),
-                        scalar2.getContentAsWideString(), bCaseSensitive, len);
+                    Cp[k] = static_cast<Nelson::logical>(compareString(elementA.getContentAsWideString(),
+                        scalar2.getContentAsWideString(), bCaseSensitive, len));
                 } else {
                     Cp[k] = false;
                 }